Cast required in printf in OSX
git-svn-id: svn://openpgp.nominet.org.uk/openpgpsdk/trunk@461 228a932f-eeec-0310-9813-fb354622df21

+++ src/advanced/adv_validate.c
@@ -93,7 +93,7 @@ validate_cb(const ops_parser_content_t *content_,ops_parse_cb_info_t *cbinfo)
 
     case OPS_PTAG_CT_USER_ATTRIBUTE:
 	assert(content->user_attribute.data.len);
-	printf("user attribute, length=%d\n",content->user_attribute.data.len);
+	printf("user attribute, length=%d\n",(int)content->user_attribute.data.len);
 	if(arg->user_attribute.data.len)
 	    ops_user_attribute_free(&arg->user_attribute);
 	arg->user_attribute=content->user_attribute;
